+++ core/src/actions_normalize.cpp
@@ -20,7 +20,6 @@
 // forward declarations
 
 icu::UnicodeString context_items_to_unicode_string(km::core::context const *context);
-km_core_usv *unicode_string_to_usv(icu::UnicodeString& src);
 
 /**
  * Normalize the output from an action to NFC, across the context | output
@@ -182,7 +181,7 @@ bool km::core::actions_normalize(
     return false;
   }
 
-  auto new_output = unicode_string_to_usv(output_nfc);
+  auto new_output = km::core::util::unicode_string_to_usv(output_nfc);
   if(!new_output) {
     // error logging handled in unicode_string_to_usv
     return false;
@@ -254,29 +253,6 @@ icu::UnicodeString context_items_to_unicode_string(km::core::context const *cont
   delete [] buf;
   return result;
 }
-
-/**
- * Helper to convert icu::UnicodeString to a UTF-32 km_core_usv buffer,
- * nul-terminated
- */
-km_core_usv *unicode_string_to_usv(icu::UnicodeString& src) {
-  UErrorCode icu_status = U_ZERO_ERROR;
-
-  km_core_usv *dst = new km_core_usv[src.length() + 1];
-
-  src.toUTF32(reinterpret_cast<UChar32*>(dst), src.length(), icu_status);
-
-  assert(U_SUCCESS(icu_status));
-  if(!U_SUCCESS(icu_status)) {
-    DebugLog("toUTF32 failed with %x", icu_status);
-    delete[] dst;
-    return nullptr;
-  }
-
-  dst[src.length()] = 0;
-  return dst;
-}
-
 
 
 /**
